backend/requirements.txt
fastapi==0.115.12
uvicorn==0.34.2
python-multipart==0.0.20
httpx==0.28.1
Pillow==11.2.1
pydantic==2.11.1
pydantic-settings==2.8.1
python-dotenv==1.1.0
pytest==8.3.5
pytest-asyncio==0.25.3
httpx==0.28.1
